# api/swagger.yaml
swagger: '2.0'
info:
  title: Enii API
  description: API documentation for the Enii backend
  version: '1.0'
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /crear_proyecto:
    post:
      summary: Crear proyecto
      tags:
        - Proyectos
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
              descripcion:
                type: string
              fecha_inicio:
                type: string
              fecha_fin:
                type: string
      responses:
        201:
          description: Proyecto creado
  /actualizar_proyecto/{project_id}:
    put:
      summary: Actualizar proyecto existente
      tags:
        - Proyectos
      parameters:
        - in: path
          name: project_id
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
              descripcion:
                type: string
              fecha_inicio:
                type: string
              fecha_fin:
                type: string
      responses:
        200:
          description: Proyecto actualizado
  /asignar_usuario_proyecto:
    patch:
      summary: Asignar usuario a proyecto
      tags:
        - Proyectos
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              proyecto_id:
                type: string
              usuario:
                type: object
              role:
                type: object
      responses:
        200:
          description: Usuario asignado
...